194403 - moved enums to V2 folder AB#194403
