+++ UKHO.ExchangeSetService.API/UKHO.ExchangeSetService.Common/Models/Enums/V2/S100ProductType.cs
@@ -1,7 +1,7 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace UKHO.ExchangeSetService.Common.Models.Enums
+namespace UKHO.ExchangeSetService.Common.Models.Enums.V2
 {
     public enum S100ProductType
     {
